+++ app/container/toolkit.py
@@ -1,35 +1,50 @@
 from dependency_injector import containers, providers
 
-from app.core.tools.map import MAPToolkitCollectors
-from app.core.tools.search import SearchToolkitCollectors
+from app.core.tools.map_api_tools import (
+    CheckPlanEligibilityTool,
+    GetContractServicesTool,
+    GetPlanSubscriptionsTool,
+)
+from app.core.tools.neo4j_search_tool import Neo4jSearchTool
 from configs import config as global_config
 
 
-# TODO : 툴킷 콘테이너 미구현 상태, 구현 필요[미구현 이유: 툴 스코프 미정]
-
-# from app.core.tools import create_agent_tools_data
-from configs import config as global_config
-
-
-# TODO : TOOLKIT 구현 필요
 class ToolkitContainer(containers.DeclarativeContainer):
 
     clients = providers.DependenciesContainer()
     db = providers.DependenciesContainer()
 
-    # _agent_tools_data = providers.Singleton(
-    #     create_agent_tools_data,
-    #     tool_collectors=providers.List(
-    #         providers.Singleton(
-    #             MAPToolkitCollectors,
-    #             map_client=clients.map_api,
-    #             method_api_key=global_config.map_method_api_keys,
-    #         ),
-    #         providers.Singleton(
-    #             SearchToolkitCollectors,
-    #             llm=clients.openai_chat_llm,
-    #             graph_db=db.neo4j_db,
-    #             graphmodel=db.neo4j_model,
-    #         ),
-    #     ),
-    # )
+    # MAP API 툴들 - 각각 올바른 method_api_key 주입
+    get_contract_services_tool = providers.Singleton(
+        GetContractServicesTool,
+        map_client=clients.map_api,
+        method_api_key=global_config.map_method_api_key_contract_mobile,
+    )
+    
+    get_plan_subscriptions_tool = providers.Singleton(
+        GetPlanSubscriptionsTool,
+        map_client=clients.map_api,
+        method_api_key=global_config.map_method_api_key_plan_add_on,
+    )
+    
+    check_plan_eligibility_tool = providers.Singleton(
+        CheckPlanEligibilityTool,
+        map_client=clients.map_api,
+        method_api_key=global_config.map_method_api_key_plan_basic,
+    )
+
+    # Neo4j Search 툴
+    neo4j_search_tool = providers.Singleton(
+        Neo4jSearchTool,
+        llm=clients.openai_chat_llm,
+        graph_db=db.neo4j_db,
+        graphmodel=db.neo4j_model,
+    )
+
+    # 활성화된 모든 툴들 (status=True인 것만)
+    all_active_tools = providers.List(
+        get_contract_services_tool,
+        get_plan_subscriptions_tool,
+        check_plan_eligibility_tool,
+        neo4j_search_tool,
+    )
