--- app/core/tools/__init__.py
@@ -1,1 +1,0 @@
-# TODO: MAP TOOL + SEARCH TOOL 합친 구조 구현 -> toolkit컨테이너 주입 필요
