+++ app/schemas/cypher.py
@@ -1,5 +1,21 @@
 from pydantic import BaseModel, Field
-from typing import Optional, Dict, Any, List, Literal
+from typing import Optional, Dict, Any, List, Literal, Union
+
+
+class Neo4jSearchRequest(BaseModel):
+    """
+    Neo4j 자연어 검색 요청 스키마
+    사용자의 자연어 질문을 받아서 Neo4j 그래프 데이터베이스 검색
+    """
+    query: str = Field(
+        ...,
+        description="자연어 검색 쿼리",
+        examples=["무제한 데이터 요금제 찾아줘", "할인 혜택이 있는 상품 보여줘"]
+    )
+    expand_search: bool = Field(
+        default=True,
+        description="검색 실패시 조건 완화 검색 수행 여부 (case1: 정확매치, case2: 조건완화)"
+    )
 
 
 class CypherRequest(BaseModel):
@@ -37,3 +53,8 @@ class CypherResponse(BaseModel):
     message: Optional[str] = Field(default=None, description="추가 메시지 (에러 메시지 등)")
     record_count: int = Field(default=0, description="반환된 레코드 수")
     database: Optional[str] = Field(default='neo4j', description="query 대상 DB")
+    
+    # Neo4j Search Tool 전용 필드들
+    case: Optional[str] = Field(default=None, description="검색 케이스 (1: 정확매치, 2: 조건완화)")
+    cypher: Optional[Union[str, List[str]]] = Field(default=None, description="실행된 Cypher 쿼리")
+    execution_time: Optional[int] = Field(default=None, description="실행 시간 (ms)")
