--- app/schemas/map/__init__.py
@@ -1,111 +1,0 @@
-from .account import (
-    AccountPeriodAmount,
-    ServicePeriodAmount,
-    UnpaidBill,
-    CurrentBillItem,
-    CurrentBillChildren,
-    FixedBillService,
-    FixedBillItem,
-)
-from .contract import (
-    ContractDetail,
-    FlexibleContractDetail,
-    PlanContractDetail,
-    TSupportContractDetail,
-    ContractPenalty2Detail,
-    RemainedContract,
-    DeviceInfo,
-    RemainingAllotmentDetail,
-    RemainingContractDetail,
-)
-from .data import (
-    DataGiftSender,
-    DataGiftLimit,
-    RegularDataGiftLimit,
-    DataGiftSendHistoryItem,
-    DataGiftSendHistory,
-    RefillCouponItem,
-)
-from .plan import (
-    AddOnProduct,
-    SupplementaryPlan,
-    AddOnSubscriptions,
-    DiscountInfo,
-    AddOnSubscriptions,
-    AddOnDetailProduct,
-    AddOnDetailSubscriptions,
-    ProductInfo,
-    PMProductInfo,
-    PMCampaignInfo,
-    BenefitInfo,
-    CustomerBenefit,
-    AddOnHistory,
-    BasicPlanCondition,
-)
-from .rate import (
-    DataUsage,
-    RecentUsage,
-    DataLimitItem,
-    DataLimit,
-    ChildDataLimitItem,
-    ChildDataLimit,
-    DataLimitChildren,
-    DataSharingService,
-    DataSharingLimit,
-    VoiceLimitItem,
-    VoiceLimit,
-    VoiceLimitChildren,
-    TFamilySharingDataLimit,
-)
-
-__all__ = [
-    "AccountPeriodAmount",
-    "ServicePeriodAmount",
-    "UnpaidBill",
-    "CurrentBillItem",
-    "CurrentBillChildren",
-    "FixedBillService",
-    "FixedBillItem",
-    "ContractDetail",
-    "FlexibleContractDetail",
-    "PlanContractDetail",
-    "TSupportContractDetail",
-    "ContractPenalty2Detail",
-    "RemainedContract",
-    "DeviceInfo",
-    "RemainingAllotmentDetail",
-    "RemainingContractDetail",
-    "DataGiftSender",
-    "DataGiftLimit",
-    "RegularDataGiftLimit",
-    "DataGiftSendHistoryItem",
-    "DataGiftSendHistory",
-    "RefillCouponItem",
-    "AddOnProduct",
-    "SupplementaryPlan",
-    "AddOnSubscriptions",
-    "DiscountInfo",
-    "AddOnSubscriptions",
-    "AddOnDetailProduct",
-    "AddOnDetailSubscriptions",
-    "ProductInfo",
-    "PMProductInfo",
-    "PMCampaignInfo",
-    "BenefitInfo",
-    "CustomerBenefit",
-    "AddOnHistory",
-    "BasicPlanCondition",
-    "DataUsage",
-    "RecentUsage",
-    "DataLimitItem",
-    "DataLimit",
-    "ChildDataLimitItem",
-    "ChildDataLimit",
-    "DataLimitChildren",
-    "DataSharingService",
-    "DataSharingLimit",
-    "VoiceLimitItem",
-    "VoiceLimit",
-    "VoiceLimitChildren",
-    "TFamilySharingDataLimit",
-]
