+++ configs/prd.py
@@ -4,3 +4,8 @@ from configs.default import BaseConfig
 class PrdConfig(BaseConfig):
     debug_mode: bool = False
     log_level: str = "ERROR"
+    
+    # MAP API keys (환경별 설정)
+    map_method_api_key_plan_add_on: str = ""
+    map_method_api_key_plan_basic: str = ""
+    map_method_api_key_contract_mobile: str = ""
